+++ packages/ai/core/generate-text/smooth-stream.test.ts
@@ -23,6 +23,24 @@ describe('smoothStream', () => {
     return Promise.resolve();
   }
 
+  describe('throws error if chunking option is invalid', async () => {
+    it('throws error if chunking strategy is invalid', async () => {
+      expect(() => {
+        smoothStream({
+          chunking: 'foo' as any,
+        });
+      }).toThrowError();
+    });
+
+    it('throws error if chunking option is null', async () => {
+      expect(() => {
+        smoothStream({
+          chunking: null as any,
+        });
+      }).toThrowError();
+    });
+  });
+
   describe('word chunking', () => {
     it('should combine partial words', async () => {
       const stream = convertArrayToReadableStream([
@@ -173,6 +191,200 @@ describe('smoothStream', () => {
         },
       ]);
     });
+
+    it('should send remaining text buffer before tool call starts', async () => {
+      const stream = convertArrayToReadableStream([
+        { type: 'text-delta', textDelta: 'I will check the' },
+        { type: 'text-delta', textDelta: ' weather in Lon' },
+        { type: 'text-delta', textDelta: 'don.' },
+        { type: 'tool-call', name: 'weather', args: { city: 'London' } },
+        { type: 'step-finish' },
+        { type: 'finish' },
+      ]).pipeThrough(
+        smoothStream({
+          delayInMs: 10,
+          _internal: { delay },
+        })({ tools: {} }),
+      );
+
+      await consumeStream(stream);
+
+      expect(events).toMatchInlineSnapshot(`
+        [
+          "delay 10",
+          {
+            "textDelta": "I ",
+            "type": "text-delta",
+          },
+          "delay 10",
+          {
+            "textDelta": "will ",
+            "type": "text-delta",
+          },
+          "delay 10",
+          {
+            "textDelta": "check ",
+            "type": "text-delta",
+          },
+          "delay 10",
+          {
+            "textDelta": "the ",
+            "type": "text-delta",
+          },
+          "delay 10",
+          {
+            "textDelta": "weather ",
+            "type": "text-delta",
+          },
+          "delay 10",
+          {
+            "textDelta": "in ",
+            "type": "text-delta",
+          },
+          {
+            "textDelta": "London.",
+            "type": "text-delta",
+          },
+          {
+            "args": {
+              "city": "London",
+            },
+            "name": "weather",
+            "type": "tool-call",
+          },
+          {
+            "type": "step-finish",
+          },
+          {
+            "type": "finish",
+          },
+        ]
+      `);
+    });
+
+    it('should send remaining text buffer before tool call starts and tool call streaming is enabled', async () => {
+      const stream = convertArrayToReadableStream([
+        { type: 'text-delta', textDelta: 'I will check the' },
+        { type: 'text-delta', textDelta: ' weather in Lon' },
+        { type: 'text-delta', textDelta: 'don.' },
+        {
+          type: 'tool-call-streaming-start',
+          name: 'weather',
+          args: { city: 'London' },
+        },
+        { type: 'tool-call-delta', name: 'weather', args: { city: 'London' } },
+        { type: 'tool-call', name: 'weather', args: { city: 'London' } },
+        { type: 'step-finish' },
+        { type: 'finish' },
+      ]).pipeThrough(
+        smoothStream({
+          delayInMs: 10,
+          _internal: { delay },
+        })({ tools: {} }),
+      );
+
+      await consumeStream(stream);
+
+      expect(events).toMatchInlineSnapshot(`
+        [
+          "delay 10",
+          {
+            "textDelta": "I ",
+            "type": "text-delta",
+          },
+          "delay 10",
+          {
+            "textDelta": "will ",
+            "type": "text-delta",
+          },
+          "delay 10",
+          {
+            "textDelta": "check ",
+            "type": "text-delta",
+          },
+          "delay 10",
+          {
+            "textDelta": "the ",
+            "type": "text-delta",
+          },
+          "delay 10",
+          {
+            "textDelta": "weather ",
+            "type": "text-delta",
+          },
+          "delay 10",
+          {
+            "textDelta": "in ",
+            "type": "text-delta",
+          },
+          {
+            "textDelta": "London.",
+            "type": "text-delta",
+          },
+          {
+            "args": {
+              "city": "London",
+            },
+            "name": "weather",
+            "type": "tool-call-streaming-start",
+          },
+          {
+            "args": {
+              "city": "London",
+            },
+            "name": "weather",
+            "type": "tool-call-delta",
+          },
+          {
+            "args": {
+              "city": "London",
+            },
+            "name": "weather",
+            "type": "tool-call",
+          },
+          {
+            "type": "step-finish",
+          },
+          {
+            "type": "finish",
+          },
+        ]
+      `);
+    });
+
+    it(`doesn't return chunks with just spaces`, async () => {
+      const stream = convertArrayToReadableStream([
+        { type: 'text-delta', textDelta: ' ' },
+        { type: 'text-delta', textDelta: ' ' },
+        { type: 'text-delta', textDelta: ' ' },
+        { type: 'text-delta', textDelta: 'foo' },
+
+        { type: 'step-finish' },
+        { type: 'finish' },
+      ]).pipeThrough(
+        smoothStream({
+          delayInMs: 10,
+          _internal: { delay },
+        })({ tools: {} }),
+      );
+
+      await consumeStream(stream);
+
+      expect(events).toMatchInlineSnapshot(`
+        [
+          {
+            "textDelta": "   foo",
+            "type": "text-delta",
+          },
+          {
+            "type": "step-finish",
+          },
+          {
+            "type": "finish",
+          },
+        ]
+      `);
+    });
   });
 
   describe('line chunking', () => {
@@ -263,6 +475,42 @@ describe('smoothStream', () => {
   });
 
   describe('custom chunking', () => {
+    it(`should return correct result for regexes that don't match from the exact start onwards`, async () => {
+      const stream = convertArrayToReadableStream([
+        { textDelta: 'Hello_, world!', type: 'text-delta' },
+        { type: 'step-finish' },
+        { type: 'finish' },
+      ]).pipeThrough(
+        smoothStream({
+          chunking: /_/,
+          delayInMs: 10,
+          _internal: { delay },
+        })({ tools: {} }),
+      );
+
+      await consumeStream(stream);
+
+      expect(events).toMatchInlineSnapshot(`
+        [
+          "delay 10",
+          {
+            "textDelta": "Hello_",
+            "type": "text-delta",
+          },
+          {
+            "textDelta": ", world!",
+            "type": "text-delta",
+          },
+          {
+            "type": "step-finish",
+          },
+          {
+            "type": "finish",
+          },
+        ]
+      `);
+    });
+
     it('should support custom chunking regexps (character-level)', async () => {
       const stream = convertArrayToReadableStream([
         { textDelta: 'Hello, world!', type: 'text-delta' },
@@ -308,6 +556,87 @@ describe('smoothStream', () => {
         { type: 'step-finish' },
         { type: 'finish' },
       ]);
+    });
+  });
+
+  describe('custom callback chunking', () => {
+    it('should support custom chunking callback', async () => {
+      const stream = convertArrayToReadableStream([
+        { textDelta: 'He_llo, ', type: 'text-delta' },
+        { textDelta: 'w_orld!', type: 'text-delta' },
+        { type: 'step-finish' },
+        { type: 'finish' },
+      ]).pipeThrough(
+        smoothStream({
+          chunking: buffer => /[^_]*_/.exec(buffer)?.[0],
+          _internal: { delay },
+        })({ tools: {} }),
+      );
+
+      await consumeStream(stream);
+
+      expect(events).toMatchInlineSnapshot(`
+        [
+          "delay 10",
+          {
+            "textDelta": "He_",
+            "type": "text-delta",
+          },
+          "delay 10",
+          {
+            "textDelta": "llo, w_",
+            "type": "text-delta",
+          },
+          {
+            "textDelta": "orld!",
+            "type": "text-delta",
+          },
+          {
+            "type": "step-finish",
+          },
+          {
+            "type": "finish",
+          },
+        ]
+      `);
+    });
+
+    describe('throws errors if the chunking function invalid matches', async () => {
+      it('throws empty match error', async () => {
+        const stream = convertArrayToReadableStream([
+          { textDelta: 'Hello, world!', type: 'text-delta' },
+          { type: 'step-finish' },
+          { type: 'finish' },
+        ]).pipeThrough(
+          smoothStream({ chunking: () => '', _internal: { delay } })({
+            tools: {},
+          }),
+        );
+
+        await expect(
+          consumeStream(stream),
+        ).rejects.toThrowErrorMatchingInlineSnapshot(
+          `[Error: Chunking function must return a non-empty string.]`,
+        );
+      });
+
+      it('throws match prefix error', async () => {
+        const stream = convertArrayToReadableStream([
+          { textDelta: 'Hello, world!', type: 'text-delta' },
+          { type: 'step-finish' },
+          { type: 'finish' },
+        ]).pipeThrough(
+          smoothStream({ chunking: () => 'world', _internal: { delay } })({
+            tools: {},
+          }),
+        );
+
+        await expect(
+          consumeStream(stream),
+        ).rejects.toThrowErrorMatchingInlineSnapshot(
+          `[Error: Chunking function must return a match that is a prefix of the buffer. Received: "world" expected to start with "Hello, world!"]`,
+        );
+      });
     });
   });