+++ packages/google-vertex/src/google-vertex-image-model.ts
@@ -3,6 +3,7 @@ import {
   Resolvable,
   combineHeaders,
   createJsonResponseHandler,
+  parseProviderOptions,
   postJsonToApi,
   resolve,
 } from '@ai-sdk/provider-utils';
@@ -65,13 +66,19 @@ export class GoogleVertexImageModel implements ImageModelV1 {
       });
     }
 
+    const vertexImageOptions = parseProviderOptions({
+      provider: 'vertex',
+      providerOptions,
+      schema: vertexImageProviderOptionsSchema,
+    });
+
     const body = {
       instances: [{ prompt }],
       parameters: {
         sampleCount: n,
         ...(aspectRatio != null ? { aspectRatio } : {}),
         ...(seed != null ? { seed } : {}),
-        ...(providerOptions.vertex ?? {}),
+        ...(vertexImageOptions ?? {}),
       },
     };
 
@@ -108,3 +115,23 @@ export class GoogleVertexImageModel implements ImageModelV1 {
 const vertexImageResponseSchema = z.object({
   predictions: z.array(z.object({ bytesBase64Encoded: z.string() })).nullish(),
 });
+
+const vertexImageProviderOptionsSchema = z.object({
+  negativePrompt: z.string().nullish(),
+  personGeneration: z
+    .enum(['dont_allow', 'allow_adult', 'allow_all'])
+    .nullish(),
+  safetySetting: z
+    .enum([
+      'block_low_and_above',
+      'block_medium_and_above',
+      'block_only_high',
+      'block_none',
+    ])
+    .nullish(),
+  addWatermark: z.boolean().nullish(),
+  storageUri: z.string().nullish(),
+});
+export type GoogleVertexImageProviderOptions = z.infer<
+  typeof vertexImageProviderOptionsSchema
+>;